pyproject.toml
[tool.poetry]
name = "data-x"
version = "0.1.0"
description = ""
authors = ["Pritesh Patel <pritesh-patel@outlook.com>"]

[tool.poetry.scripts]
data-x = "data_x.main:app"

[tool.poetry.dependencies]
python = ">=3.8,<=3.10"
typer = "^0.3.2"
pyarrow = "^4.0.1"
PyYAML = "^5.4.1"
marshmallow = "^3.12.2"
dataprep = "^0.3.0"
pandas = "^1.3.0"
dask = {extras = ["distributed"], version = "^2.30.0"}
modin = {extras = ["dask"], version = "^0.10.1"}
pandas-profiling = "^3.0.0"
simpletransformers = "^0.61.13"
torch = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"